+++ PrepperBox.Dto/RequestMessages/CreateProductFamilyRequest.cs
@@ -7,5 +7,5 @@ public sealed record CreateProductFamilyRequest(
     CategoryRef CategoryId,
     string Name,
     UnitOfMeasure UnitOfMeasure,
-    int MinimumStockLevel
+    decimal MinimumStockLevel
 );
